refactor: move ui state into ui module
