+++ rigflow-client/src/app/mod.rs
@@ -2,7 +2,7 @@ pub mod bands;
 pub mod frequency_view;
 pub mod layout;
 pub mod om_bands;
-pub mod state;
+//pub mod state;
 pub mod stats;
 pub mod spectrum_utils;
 pub mod waterfall;
