+++ rigflow-client/src/ui/app.rs
@@ -7,10 +7,12 @@ use crate::app::layout::{
     LEFT_GUTTER, RIGHT_GUTTER, WATERFALL_IMAGE_HEIGHT, WATERFALL_IMAGE_WIDTH,
 };
 use crate::app::om_bands::LicenseClass;
-use crate::app::state::UiState;
 use crate::net::control::ControlCommand;
-use crate::ui::spectrum_view::{
-    draw_spectrum_plot, x_frac_to_frequency_hz, SpectrumInteraction,
+use crate::ui::{
+    spectrum_view::{
+	draw_spectrum_plot, x_frac_to_frequency_hz, SpectrumInteraction,
+    },
+    state::UiState,
 };
 
 pub struct RigflowApp {
