+++ rigflow-client/src/ui/mod.rs
@@ -4,7 +4,7 @@ pub mod app;
 //pub mod layout;
 //pub mod om_bands;
 //pub mod spectrum_utils;
-//pub mod state;
+pub mod state;
 //pub mod stats;
 //pub mod waterfall;
 pub mod spectrum_view;
